requirements.txt
Flask
Pillow
fpdf
python-dotenv
